LPS-58632 Check collisions when adding AAAA but a AAAA.txt exists and vice versa

+++ portal-impl/src/com/liferay/portlet/documentlibrary/lar/FileEntryStagedModelDataHandler.java
@@ -361,25 +361,22 @@ public class FileEntryStagedModelDataHandler
 					portletDataContext.getScopeGroupId(), folderId,
 					fileEntryTitle);
 
-				if (existingTitleFileEntry != null) {
-					if ((fileEntry.getGroupId() ==
-							portletDataContext.getSourceGroupId()) &&
-						portletDataContext.
-							isDataStrategyMirrorWithOverwriting()) {
+				if ((existingTitleFileEntry != null) &&
+					(fileEntry.getGroupId() ==
+						portletDataContext.getSourceGroupId()) &&
+					portletDataContext.isDataStrategyMirrorWithOverwriting()) {
 
-						DLAppLocalServiceUtil.deleteFileEntry(
-							existingTitleFileEntry.getFileEntryId());
-					}
-					else {
-						fileEntryTitle = DLFileEntryLocalServiceUtil.getUniqueTitle(
-								portletDataContext.getScopeGroupId(), folderId,
-								-1, fileEntryTitle, fileEntry.getExtension());
-					}
+					DLAppLocalServiceUtil.deleteFileEntry(
+						existingTitleFileEntry.getFileEntryId());
 				}
 
 				serviceContext.setAttribute(
 					"fileVersionUuid", fileVersion.getUuid());
 				serviceContext.setUuid(fileEntry.getUuid());
+
+				fileEntryTitle = DLFileEntryLocalServiceUtil.getUniqueTitle(
+						portletDataContext.getScopeGroupId(), folderId, -1,
+						fileEntryTitle, fileEntry.getExtension());
 
 				importedFileEntry = DLAppLocalServiceUtil.addFileEntry(
 					userId, portletDataContext.getScopeGroupId(), folderId,
